+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added new EC2 Tag COMPONENT_INIT_STATUS
+
 ### Changed
 - Change default orchestrator and replicator credentials to overwrite-me/overwrite-me
+- Changed process of determining the healthy publish instance
 
 ## [2.0.0] - unknown